Renamed project folder to remove left over name from move to .net core.
Removed database logging since it wasn't used and just adding bloat and filespace usage.
