+++ GlucoseTray.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28803.202
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GlucoseTrayCore", "GlucoseTrayCore\GlucoseTrayCore.csproj", "{2778CBCC-020C-4F3D-99BC-E8C231FB7F53}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GlucoseTray", "GlucoseTray\GlucoseTray.csproj", "{0A0B15BE-CA2F-4B32-A338-C8B2DF04060E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2778CBCC-020C-4F3D-99BC-E8C231FB7F53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2778CBCC-020C-4F3D-99BC-E8C231FB7F53}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2778CBCC-020C-4F3D-99BC-E8C231FB7F53}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2778CBCC-020C-4F3D-99BC-E8C231FB7F53}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0A0B15BE-CA2F-4B32-A338-C8B2DF04060E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0A0B15BE-CA2F-4B32-A338-C8B2DF04060E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0A0B15BE-CA2F-4B32-A338-C8B2DF04060E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0A0B15BE-CA2F-4B32-A338-C8B2DF04060E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
