+++ GlucoseTray/Enums/DexcomServer.cs
