+++ GlucoseTray/Enums/FetchMethod.cs
