+++ GlucoseTray/Enums/TrendResult.cs
