+++ GlucoseTray/Enums/UpDown.cs
