+++ GlucoseTray/Extensions/GlucoseFetchResultExtensions.cs
