+++ GlucoseTray/Extensions/StringExtensions.cs
