+++ GlucoseTray/GlucoseTray.csproj
@@ -6,7 +6,6 @@
     <RootNamespace>GlucoseTrayCore</RootNamespace>
     <PublishReadyToRun>true</PublishReadyToRun>
     <PublishSingleFile>true</PublishSingleFile>
-    <PublishTrimmed>true</PublishTrimmed>
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
     <ApplicationIcon />
     <StartupObject />
