+++ GlucoseTray/GlucoseTraySettings.cs
