+++ GlucoseTray/Models/DexcomResult.cs
