+++ GlucoseTray/Models/GlucoseResult.cs
