+++ GlucoseTray/Models/NightScoutResult.cs
