+++ GlucoseTray/Models/NightScoutStatus.cs
