+++ GlucoseTray/Models/NightScoutStatusResult.cs
