+++ GlucoseTray/Services/IconService.cs
