+++ GlucoseTray/Services/SettingsService.cs
