+++ GlucoseTray/Services/StringEncryptionService.cs
