+++ GlucoseTray/Views/SettingsWindow.xaml
@@ -67,11 +67,6 @@
         <TextBox TabIndex="18" x:Name="txt_polling" Grid.Column="4" HorizontalAlignment="Left" Grid.Row="7" Text="{Binding PollingThreshold}" TextWrapping="Wrap" Margin="5,4,0,3" Width="50"/>
         <Label Content="Results Stale After (minutes)" Grid.Column="3" HorizontalAlignment="Center" Grid.Row="8" VerticalAlignment="Center" Height="26" Width="157"/>
         <TextBox TabIndex="19" x:Name="txt_stale" Grid.Column="4" HorizontalAlignment="Left" Grid.Row="8" Text="{Binding StaleResultsThreshold}" TextWrapping="Wrap" Margin="5,4,0,3" Width="50"/>
-        <Label Content="Log Level" Grid.Column="3" HorizontalAlignment="Left" Grid.Row="9" VerticalAlignment="Center" Grid.ColumnSpan="2" Height="26" Margin="97,0,0,0" Width="60"/>
-        <ComboBox TabIndex="20" x:Name="combobox_loglevel" Grid.Column="4" HorizontalAlignment="Left" Grid.Row="9" VerticalAlignment="Center" Width="120" Margin="5,0,0,0" SelectedIndex="0" Height="22"/>
-        <CheckBox TabIndex="21" x:Name="checkbox_debug" Content="Debug Mode" Grid.Column="4" HorizontalAlignment="Left" Grid.Row="10" VerticalAlignment="Center" Margin="5,0,0,0" IsChecked="{Binding EnableDebugMode}" Height="15" Width="90"/>
-        <Label Content="Database Location" Grid.Column="3" HorizontalAlignment="Left" Grid.Row="11" VerticalAlignment="Center" Grid.ColumnSpan="2" Height="26" Margin="50,0,0,0" Width="107"/>
-        <TextBox TabIndex="22" x:Name="txt_db_location" Grid.Column="4" Grid.Row="11" Text="{Binding DatabaseLocation}" TextWrapping="Wrap" Margin="5,4,0,3"/>
         <Button TabIndex="23" x:Name="button_save" Content="Save" Grid.Column="2" Grid.Row="13" VerticalAlignment="Center" Height="35" Click="Button_Save_Click" Grid.ColumnSpan="2" Margin="25,0,0,0"/>
         <CheckBox TabIndex="13" x:Name="checkbox_alert_high" IsChecked="{Binding HighAlert}" Content="" Grid.Column="2" HorizontalAlignment="Center" Grid.Row="7" VerticalAlignment="Center"/>
         <CheckBox TabIndex="14" x:Name="checkbox_alert_warning_high" IsChecked="{Binding WarningHighAlert}" Content="" Grid.Column="2" HorizontalAlignment="Center" Grid.Row="8" VerticalAlignment="Center"/>
