+++ GlucoseTray/app.manifest
