+++ GlucoseTray/appsettings.json
